ai_engine/requirements.txt
# SYNAPSE AI Engine — pinned requirements
# All versions pinned to compatible set (tested together)

# ── FastAPI stack ──────────────────────────────────────────────
fastapi>=0.111,<0.112
uvicorn[standard]>=0.29,<0.30
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<0.1

# ── Rate limiting ──────────────────────────────────────────────
slowapi>=0.1.9,<0.2

# ── AI / ML — pinned to compatible set ────────────────────────
# langchain-google-genai 2.x works with langchain-core>=0.3
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.2,<0.3
langchain-google-genai>=2.0,<3.0
langchain-anthropic>=0.3,<0.4      # TASK-302: Anthropic Claude support
langchain-ollama>=0.2,<0.4          # TASK-302: Ollama local LLM support
tavily-python>=0.5,<1.0             # TASK-303-B1: Tavily web search tool
pymupdf>=1.24,<2.0                  # TASK-303-B3: PDF document reader (fitz)
matplotlib>=3.9,<4.0                # TASK-303-B4: Chart/visualization generator
langgraph>=0.2,<0.3
langsmith>=0.2,<0.3
sentence-transformers>=3.0,<4.0
torch>=2.3,<3.0

# ── Database ───────────────────────────────────────────────────
psycopg2-binary>=2.9,<3.0
pgvector>=0.2,<1.0
SQLAlchemy>=2.0,<3.0

# ── Cache ──────────────────────────────────────────────────────
redis>=5.0,<6.0

# ── Monitoring ─────────────────────────────────────────────────
structlog>=24.2,<25.0
sentry-sdk[fastapi]>=2.5,<3.0

# ── OpenTelemetry (TASK-504-B2) — optional, disabled by default ────────────────
# Uncomment to enable distributed tracing via OTEL_ENABLED=true
# opentelemetry-api>=1.20,<2.0
# opentelemetry-sdk>=1.20,<2.0
# opentelemetry-instrumentation>=0.41,<1.0
# opentelemetry-instrumentation-fastapi>=0.41,<1.0
# opentelemetry-exporter-otlp>=0.41,<1.0

# ── Token estimation (TASK-004-B3) ─────────────────────────────────────────────
tiktoken>=0.7,<1.0

# ── PII detection — optional, graceful degradation if not installed ────────────
# presidio-analyzer>=2.2,<3.0
# presidio-anonymizer>=2.2,<3.0

# ── Utilities ──────────────────────────────────────────────────
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
